requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.1
djangorestframework_simplejwt==5.2.2
django_phonenumber_field==7.0.1
geopy==2.3.0
phonenumbers==8.13.10
Pillow==9.5.0
drf_yasg==1.21.5
python_dotenv==1.0.0
flake8==6.0.0
django_cors_headers==3.14.0
google_auth==2.17.3
dialogflow==2.0.0
daphne==4.0.0
channels==4.0.0
stripe==6.0.0
elasticsearch==8.10.0
elasticsearch_dsl==8.9.0

# packages throws error while seting up project
psycopg2==2.9.6
